Add the answer formatter
This commit will implement the answer formatter. This is a JS function
which formats the 'answers' section of the questionaire into a new
object with user fiednly answers suitable to display on the check your
answers page.

The formatter serves to implement the following rules:

- All individual answers are formatted in a readable format, i.e. capitalised,
appropriately spaced, with true/false converted to Yes, No
- Questions with more than 1 but less than 5 fields are merged onto a
single line, separated by spaces.
- Answers to boolean questions are separated with a comma, not a space
- Questions with more than 4 fields are formatted into a multiline
string, separated with line breaks.
y