Change scope of answer object sent to transformer
Previous to this commit, the entire answer object was being passed to the
transformer on each request. This was preventing fields from
re-populating previously answered questions.

This commit will change the scope of the answer object so that each
transformation only receives the relevent answer.

This commit also fixes some minor bugs.
 - Quotations around the back buttons URL have been added.
 - Questionaire object is no longer populated with test data
 - package.json no longer points to a local copy of the transformer