Implement type key
This commit makes use of the 'type' key which is now included in the
questionnaire object.

Having this key available allows for the nunjucks string to be altered
and to send a response suitable for the type of page being routed to.

Moved 'start' and 'summary' up a level in the routes json object

Updated questionnaire

Fixed date display bug