Revise conditional reveal validation
This commit will add code which identifies answers to conditionally
revealing questions which are not selected, and removes them. This
keeps the answer object free of answers which we are not interested
in, and resolves a bug where these additional answers were causing
the schema to fail validation.

This commit also updates the questionnaire to an up to date version.
This update includes a fix to a bug which results in hidden fields
throwing a required error despite their parent control not being
used.

This commit also includes an updated questionnaie to be compatible
with the above changes.