Update uiSchema
This commit updates the uiSchema to include schanges to the CYA page.

This commit facilitates the use of the "buttonText" key to change
the default button text on a section.

This commit also checks whether a page is a summary page or not
using the uiSchema.

This commit also includes a change to how a page redirects when a
query string is used. The variable is now called 'next'.

Finally, this commit also fixes two broken tests.