Update validation (#33)
The commit adds the 'novalidate' property to the form. This will
disable the browser's default validation to allow the schema to
validate the users email address.

This commit also updates the regex used to test phone numbers to
be more efficient.

Finally, this commit also adds the correct link to the "another
relevant place" link on the start page.