Add displayBackButton to UISchema
This commit adds the ability to show or hide the back button on any
page using the uischema. By default, this value will always be true,
the back button will only be hidden if this variable is explicitly
set to false.

This commit also removes some redundant functions from the form
helper.